// src/agenda/enums/agendaErrors.enum.ts
export enum AgendaErrorsEnum {
  APPOINTMENT_NOT_RELATED_TO_AGENDA = 'APPOINTMENT_NOT_RELATED_TO_AGENDA',
  APPOINTMENT_ACCOMPLISHED = 'APPOINTMENT_ACCOMPLISHED',
  BAD_DATE_INPUT = 'BAD_DATE_INPUT',
  COACH_HAS_NO_AVAILABILITY = 'COACH_HAS_NO_AVAILABILITY',
  EXCEEDS_MAXIMUN_APPOINTMENT_DISTANCE = 'EXCEEDS_MAXIMUN_APPOINTMENT_DISTANCE',
  EXCEEDS_MAXIMUN_APPOINTMENT_PER_MONTH = 'EXCEEDS_MAXIMUN_APPOINTMENT_PER_MONTH',
  LESS_THAN_MINIMUN_SESSION_DURATION = 'LESS_THAN_MINIMUN_SESSION_DURATION',
}